增加 JVM下guava和caffeine实现的缓存

+++ src/main/kotlin/io/github/rcarlosdasilva/weixin/handler/Cache.kt
@@ -100,7 +100,7 @@ interface CacheStorage<V : Cacheable> {
   /**
    * 缓存大小
    */
-  fun size(): Int
+  fun size(): Long
 
   /**
    * 清空
